make arrays have consistent types
in order to call into pybind APIs without casting by the user.
This involves changing most array outputs to double
precision (wavelength in particular needs to be double, and
then making most things double is convenient), and casting all
arrays to consistent types on read so the user doesn't have to
(and to avoid byte swapping problems).

+++ python/pfs/datamodel/pfsFiberArray.py
@@ -102,10 +102,10 @@ class PfsFiberArray(PfsSimpleSpectrum):
             Keyword arguments for constructing spectrum.
         """
         data = super()._readImpl(fits)
-        data["sky"] = fits["SKY"].data
+        data["sky"] = fits["SKY"].data.astype(float)
         data["observations"] = Observations.fromFits(fits)
-        data["covar"] = fits["COVAR"].data
-        data["covar2"] = fits["COVAR2"].data
+        data["covar"] = fits["COVAR"].data.astype(float)
+        data["covar2"] = fits["COVAR2"].data.astype(float)
         try:
             fluxTable = FluxTable.fromFits(fits)
         except KeyError as exc:
